moving funcs to utils
